Bootstrap testing with tox
Tox is a great tool to allow developers to run local testing in a
sandbox environment.  We use this pattern with other python project in
ansible.

We also add flake8 support to help format our code. For now, we disable
a lot of things in flake8 and future PRs will correct them.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>