Initial commit for container support
We eventually want to support ansible-catalog via a container, this is
the first step to do that. We can leverage the existing python-builder /
python-base images to create the container since this is a python
project.

Future improvements will be do some functional testing using the
container to confirm everything is properly configured. So far, this
just builds and installs the project inside the container.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>