Consolidate codebase in a single `ansible_catalog` package
By default Django creates apps as a top level packages in the project
directory as django apps were usually deployed as project directory
and are not designed for packaging.

When building a distribution package, top level packages may cause
conflicts with other packages installed at the same time in the system
or virtual environmnet.

When packaging a python application or a library, it is common practice
to use a single top-level import package (usually but not necessarily
named after the distribution name).

This patch moves `common` and `main` top-level packages into the
`ansible_catalog` namespace.