Remove Basic Authentication
With Basic authentication enabled when we fail to refresh a token
from keycloak we send back a 401. The 401 response contains a header
WWW-Authenticate: Basic realm="api"

When the Web browser sees this header it pops up the Login dialog.
If we remove the Basic Authentication this header is not set.

https://issues.redhat.com/browse/SSP-2708