Added support for kuberenetes job to do keycloak setup
For the minikube dev env we were running the keycloak setup everytime
we were doing redeployment of the app. It is unnecessary to do it
everytime it just needs to be done once. The app will start a lot
quicker now for developers as they test their changes. With this PR
we will run it once as a kubernetes job.