[SSP-2922] Once anchored to a source dont allow changes
Once an On Prem Pinakes is connected to a Source (Automation
Controller) we cannot have it connect to another Source.
If the user tries to change the Source report a user friendly
localized message. The logic is based on the install_uuid from
the Automation Controller once docked to a controller we store
the install_uuid for the Source.