For minikube use environment vars from a single file
To ease customization we should be able to set all the env vars
in a single file and using envFrom feature in minikube we can create
all the environment variables in the pods. Previously we had to
create a separte configMap for every setting which makes customization
difficult.
The file name is called

* minikube_env_vars

A sample file is located in ./tools/minikube/env_vars.sample
A user can copy this file to the root directory as minikube_env_vars

To ease the transition to this logic the start_pods.sh will create
this file from the earlier required env vars